requirements.txt
pytest==8.2.0
pytest-playwright==0.5.0
playwright==1.44.0
pytest-html==4.1.1
pymysql==1.1.1
python-dotenv==1.0.1